feat: module separation, feat: recent events in report
