+++ requirements.txt
@@ -4,7 +4,7 @@ gitdb==4.0.11
 GitPython==3.1.43
 idna==3.7
 packaging==24.1
-python-dateutil
+python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
 requests==2.32.3
 setuptools==69.2.0
