+++ src/gh_fake_analyzer/__init__.py
@@ -1,4 +1,3 @@
-from .analyze import run_analyzer
-from .monitor import run_monitor
+from .terminal import start_terminal
 
-__all__ = ["run_analyzer", "run_monitor"]
+__all__ = ["start_terminal"]
