+++ src/gh_fake_analyzer/modules/report.py
@@ -1,0 +1,1 @@
+# add generate_report here
